+++ test/Journalist.EventStore.IntegrationTests/Streams/EventStreamsTests.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Linq;
 using System.Threading.Tasks;
+using Journalist.EventStore.Connection;
 using Journalist.EventStore.Events;
 using Journalist.EventStore.Streams;
 using Journalist.EventStore.IntegrationTests.Infrastructure.TestData;